git: Expand `withGitStore`
When initializing a new instance of anything TreeDb-like, this must be done with
a preexisting (usually hard-coded) `BlockHeader`. However, since a prepopulated
Git store can live on disk just fine, we only need a filepath in order to
reconnect to it.

This creates a kind of configuration combinatorics, where it could be that both
a genesis block were given, and yet the indicated filepath already exists (and
contains something). In this case, the entire process should fail as being
"misconfigured".

The bottom line is that any initialized `GitStore` (or anything TreeDB-like)
have contain _at least_ its genesis block.