[on-disk cas] remove "emptyCas" from Cas typeclass and fixup callers.
emptyCas :: IO (cas a) as a member of the typeclass is mistyped here -- for an
on-disk store you don't have enough information to proceed (your creation
function will look like "FilePath -> IO (DB a)").

After removing it we can proceed with adapting the on-disk CAS I wrote a while
ago to Lars' typeclass scheme.