Pact service: recover from fatal tx error in newBlock. (#831)
* Pact service: recover from fatal tx error in newBlock.

Sometimes we run the pre-block check on a set of transactions and it reports
that they all have enough gas to be playable. However, sometimes a transaction
in the middle of the block can cause a subsequent transaction to fail the
buy-gas check. In this case, we were getting caught in a fatal loop, since the
offending txs keep getting re-selected instead of being bad-listed as intended.

Fixes #829.

Changes:

  1. add a mempoolAddToBadList function to Mempool

  2. propagate it to PactService's MemPoolAccess

  3. add a new exception constructor to PactException that contains the failing
     tx hash, and one for coinbase exception

  4. modify TransactionExec to throw the new exception constructor on fatal
     error instead of "Internal error", and fixup the other caller to "fatal"
     to throw the coinbase exception instead

  5. catch the new constructor in `execNewBlock` and use it to bad-list the bad
     transaction so it isn't re-selected.

  6. test for new mempoolAddToBadList

  7. test for badlisting behavior inside PactService.

* Fix bad merge