Create "chainweb-portable" package, which can run on other linuxes
This has not been extensively tested, but it did work on the one Ubuntu machine I tried.

To build, do this: nix-build ./project.nix -A proj.ghc.chainweb-portable

To install, take all the files in the output and copy them to /opt/chainweb on the target machine.  This prefix directory can be changed by supplying a different argument in `project.nix`, but it cannot be made fully relocatable easily.  Perhaps something like [rtldi](http://bitwagon.com/rtldi/rtldi.html) could help, but I haven't evaluated it.  It would also be possible to supply a script based on `patchelf` to change the prefix without rebuilding.