Build a proper default.nix with a modern nixpkgs pin
This default.nix file uses Nix release 22.05 (May 2022) as its basis, with GHC
8.10.7 as the compiler, and setting version pins where necessary for
compatibility with chainweb.cabal. It strives to keep overrides to a minimum,
and to avoid disabling as many build-time checks as possible.

Running nix-build in the chainweb-node directory should produce a build that's
very close to our Linux-based production build.