Fix nix builds by bumping hackage (#1667)
This PR is an alternative to #1664. The main difference is that we bump the `hackage` snapshot by defining it as a direct input of our flake and telling `haskellNix` to `follow` it. The we can just `nix flake lock --update-input hackage` to bump the snapshot without actually bumping `haskell.nix`. I think this has 2 main advantages:
* Bumping `haskell.nix` may introduce unrelated changes to the build infrastructure. Changes one might not be willing to deal with just to bump the hackage snapshot.
* Bumping `haskell.nix` triggers a deeper transitive dependency rebuild session compared to just bumping `hackage`, which only requires a rebuild of the project's Haskell package dependencies. This can mean a 10-hour rebuild vs. a 2-hour rebuild.