Replace legacy default.nix with flake based default.nix (#1670)
This PR removes the legacy Nix infrastructure for building the project. The legacy infra was based on `haskell4nix` from `nixpkgs`, but our recently used `flake.nix` uses `haskell.nix` and removes most of the Nix maintenance burden. The legacy infra has been broken since the 2.19.1 release, so it's been causing trouble for users that tried to build the project with the legacy `nix-build` command, instead of `nix build`.

This PR also removes `docker.nix` that relied on the legacy infra. We'll reintroduce Nix-built docker images in a subsequent PR using the new Nix infra.

* Remove the legacy nix infra

* Rename default-flake.nix to default.nix