service: return correct tx hash in txindex inner join (#1775)
Summary: The INNER JOIN used to return the result set when looking up a
transaction hash was `zip`'d without taking into account the ordering of the
result set. This in turn meant that tx ID's got mapped to the wrong block IDs,
resulting in lookup failures.

This could be reproduced with the following minimal `/poll` query with only two
request keys:

    {"requestKeys": [
      "1zirlzrAQi4Zlzew5t8LF1jTtkom6YZS9qT-K9IEzHI",
      "MdYI2HrYztPbpnAN5BXCT4IRw10XqcIU1qAPLC0l9Eo"
    ]}

which would return the empty JSON object.

Aside from all of that, the error handling for `internalPoll` is rather poor as
it exists purely within `MaybeT`, giving no chance of reporting any meaningful
errors. (I threw away a small cleanup for this since it was too debugging
oriented, and needs to go through another round of review.)

Closes #1732.


Change-Id: I3e6f1a98a5d867125c54cac8d47d01ba

Signed-off-by: Austin Seipp <aseipp@pobox.com>