Expose pact without relying on path caching
Trying to expose the `src` of the pact input by caching it as a nix
store path proved to be fragile. It's possible to output such paths
from recursive-nix derivations, but those paths will be very tricky to
access by downstream Nix expressions.

This  commit instead uses pure metadata to reconstruct the pact src
definition at the recursive layer.