poll endpoint: refactorings (#1790)
* service: refactor internal /poll body a bit

Summary: Just to make things more readable.

Signed-off-by: Austin Seipp <aseipp@pobox.com>
Change-Id: Icb9b8ac03473b6c637c1ec6dc88b5c15

* pact/restapi: migrate some use of `MaybeT` to `ExceptT`

Summary: `MaybeT` is alluring and easy to use, but destroys debuggability in
practice unless used with extreme care. Do not use it inside `internalPoll` in
order to ease debugging; the hope is that this will make bugs like #1732 easier
to find in the future.

Signed-off-by: Austin Seipp <aseipp@pobox.com>
Change-Id: I93f98d0213b48c9af809620b72d92d08

---------

Signed-off-by: Austin Seipp <aseipp@pobox.com>