Use checkpointer rewind in compaction
Compaction to prior block heights should be equivalent to rewinding and then
compacting; this change makes that the implementation more or less, by adding
a rewind to the target before compaction of user tables. Note that it doesn't
change the compaction query itself, which is still compacting "to a height",
despite not really being necessary.

Change-Id: I939e1aa885175ae8db1de37455a8a0b51aae57ee