feature: newblock resumes blocks in progress
This adds a new pact request called ContinueBlock to do this resumption
and also adds an option to NewBlock which allows creating empty blocks;
as well, the state involved in blocks in progress is encapsulated in a 
new BlockInProgress type.

Test plan is a new test that checks that a block created via NewBlock
is identical to a block created via a combination of NewBlock and
ContinueBlock.

Change-Id: Id0e6f9ac29cbef4e1b029db5be7aee0b80c2f9c8