log: disable telemetry backend by default
Just to be clear for any readers: "telemetry" is structured logging for node
operators to use to monitor their nodes. It is not sent to Kadena. By default,
it is sent to stdout, and it can be configured to be sent to an Elasticsearch
instance of the user's choosing.

This commit disables even logging telemetry to the terminal by default, because
usually it's very verbose and not useful.

Change-Id: I83ed57c79fa99151909c76d7de70f6edc88106f8