fix PactExec tests
The problem here is that PactExec used to be able to set the gas model
to freeGasModel. Now, PactService doesn't support configuring the gas
model (for good reason), so it started charging gas; 16 gas per
coin.get-balance call, and 1 gas per function application otherwise.

This doesn't imply a breaking change in Pact.