fix mempool refill test
This test began failing because the mempool became more strict in
requiring that its transactions are validated by the user validation
function passed in. The test aims to establish that duplicate
transactions cannot make it into blocks, so it suffices to change the
test to stop failing when the transaction is invalid (a normal case
for a duplicate) and instead fail when the subsequent block manages to
include the duplicate transaction.