fix block gas tests
The block gas tests were actually working, but they were mysterious. I
have made them less mysterious by establishing that in fact block
validation failures are a "success" case when the tx fits in the
block, because if the tx did not fit in the block, the block gas limit
error would be raised before the validation failure, and the
validation failure is otherwise expected, because the block is not
valid in the first place, seeing as we have to circumvent newBlock to
even construct blocks for testing.