Remove redundant `force` calls in updatePeerDb
This code is already strict enough. ixset-typed is a set and a strict
list of maps. When updated, ixset-typed will force the set and maps
to whnf, and insert the new element into a Set. Sets are strict in 
their elements and maps are strict in their keys, and the value type
`PeerEntry` is itself strict in its components.