randomized peer ordering
We randomize the peer ordering by assigning a peer a random number
that is used subsequently to order it with respect to other peers upon
request from the /peer GET endpoint.

An alternative of generating a seed on startup that would be used to
"salt" the peer ordering was considered but ran into implementation
difficulties and seems equivalent.