optimize RemotePactTest and add new invalid tx test
This test is optimized by sharing the TLS cert and HTTP manager
between tests. Otherwise the rest of the testing fixture is *not*
shared.

This PR also adds some tasty steps to a few tests, which makes it
clearer what they're spending their time on.

There's one extra invalid tx test as well, testing the error from
sending to the wrong chain.

Change-Id: Id00000004e9b814aa79622e2332683c496cda3cc