Make versionquirks testable
Prior, we'd say "the command with request key X has gas Y". Now we say
"the command on chain X with block height H and at index I has gas Y".
This makes it possible to test without tying the knot anywhere,
because the request key is a hash that includes the version name, and
also lets us integrate it into Pact 5 later.

Change-Id: Id00000007bbb29b009a2b2e4a0cf36180c8df1a7