Fix broken verifier gassing in Pact 5
This is a tricky thing to get right. The tests betrayed that we had it
wrong, but the correct thing to do here was hard to find in the
presence of both saturating unsigned arithmetic (which doesn't seem to
fully work, I got overflow errors when subtraction went below 0) and
milligas/gas conversion/rounding.

Change-Id: Id00000006858d16ed26102355a3fe9b80e6c8b3a