Stop double- encoding and decoding in the mempool
Prior, when we fetch a tx from a remote mempool, we decode it, then
re-encode it to insert it in the local mempool. This deletes the
re-encoding.

Prior, when we send a tx to a remote mempool, we decode it, deepseq
it, then encode the decoded tx to send to the remote mempool. This
deletes the deepseq and the re-decoding. This is probably a big deal.

Prior, when a tx was inserted into the local mempool via the mempool
insert endpoint, we would decode it in the handler, then re-encode it
to insert it in the local mempool. This deletes the re-encoding.

Change-Id: Id0000000adb067b5766554c94d8188a2b804cd9b