BlockHistoryMigration resilience to consensus out of sync with Pact
Sometimes consensus and Pact are out of sync; this is normal and
expected because catchup is not atomic w.r.t. payload providers. In
particular Pact can be ahead of consensus.

However, consensus' store of validated block headers is used to migrate
the BlockHistory table. If the corresponding headers are missing from
consensus, we can't migrate those rows of the BlockHistory table. Thus
*we need to avoid erroring out in that case*.

Additionally, the rows of all other Pact tables should agree with the
highest BlockHistory row. Thus *we need to execute a rewind to that block
at the end of the migration*.

Test Plan:
replay on mainnet database.

Change-Id: Id000000074e2d41ff0d0254f3d31e9c712adc6e5

Reviewers: jmcardon, larskuhtz, rsoeldner

Reviewed By: rsoeldner, larskuhtz

Pull Request: https://github.com/kadena-io/chainweb-node/pull/2261