Change the compaction command line interface
1. Fixes a bug in the command line flags.

The flag parsers did not match the constructor fields.

2. Makes only compacting sqlite the default.
3. Makes keeping TransactionIndex the default.

This way the default is "safe" and does not lose any data any user might
want. They can compact rocksdb and TransactionIndex in addition if they
want, but that should be a conscious choice, so that their API still
serves all relevant data.